+++ src/main/java/com/supermarket/kata/code_supermarket_pricing/UnitPrice.java
@@ -1,8 +1,10 @@
 package com.supermarket.kata.code_supermarket_pricing;
 
+import java.math.BigDecimal;
+
 public class UnitPrice extends Price {
 
-	public UnitPrice(Float amout) {
+	public UnitPrice(BigDecimal amout) {
 		super(amout);
 
 	}
